# docs/swagger.yaml
basePath: /
definitions:
...
  common.APIError:
    properties:
      code:
        type: string
      context:
        additionalProperties: true
        type: object
      details:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  common.APIResponse:
    properties:
      data: {}
      error:
        $ref: "#/definitions/common.APIError"
      message:
        type: string
      success:
        type: boolean
    type: object
...
  marketplace.ActorType:
    enum:
      - USER
      - SYSTEM
      - ADMIN
    type: string
    x-enum-varnames:
      - ActorTypeUser
      - ActorTypeSystem
      - ActorTypeAdmin
...
  marketplace.AuditEventResponse:
    properties:
      actor_id:
        example: USER_456
        type: string
      actor_type:
        allOf:
          - $ref: "#/definitions/marketplace.ActorType"
        example: USER
      event_id:
        example: EVT_1234567890
        type: string
      event_type:
        allOf:
          - $ref: "#/definitions/marketplace.AuctionEventType"
        example: BID_PLACED
      id:
        example: uuid-123
        type: string
      listing_id:
        example: LST_1234567890
        type: string
      summary:
        example: Bid placed on listing LST_1234567890 by USER_456
        type: string
      timestamp:
        example: "2024-01-14T10:30:00Z"
        type: string
    type: object
  marketplace.AuditLogResponse:
    properties:
      events:
        items:
          $ref: "#/definitions/marketplace.AuditEventResponse"
        type: array
      pagination:
        $ref: "#/definitions/common.PaginationMeta"
    type: object
...
  marketplace.ForceCloseListingRequest:
    properties:
      reason:
        example: Policy violation - fraudulent listing
        type: string
    required:
      - reason
    type: object
  marketplace.ListingListResponse:
    properties:
      listings:
        items:
          $ref: "#/definitions/marketplace.ListingResponse"
        type: array
      pagination:
        $ref: "#/definitions/common.PaginationMeta"
    type: object
...
  marketplace.MarketplaceStatsResponse:
    properties:
      active_listings:
        example: 450
        type: integer
      average_bids_per_listing:
        example: 4.5
        type: number
      average_listing_value:
        example: "2000.00"
        type: string
      closed_listings:
        example: 600
        type: integer
      expired_listings:
        example: 200
        type: integer
      total_bids:
        example: 5670
        type: integer
      total_listings:
        example: 1250
        type: integer
      total_value:
        example: "2500000.00"
        type: string
    type: object
  marketplace.RemoveBidRequest:
    properties:
      reason:
        example: Fraudulent bid detected
        type: string
    required:
      - reason
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the KisanLink E-commerce service API.
  title: KisanLink E-commerce API
  version: "1.0"
paths:
  /admin/v1/marketplace/audit-log:
    get:
      consumes:
        - application/json
      description:
        Retrieve audit log of marketplace activities for compliance and
        monitoring
      parameters:
        - description: Filter by listing ID
          in: query
          name: listing_id
          type: string
        - description: Filter by event type
          enum:
            - LISTING_CREATED
            - BID_PLACED
            - BID_OUTBID
            - LISTING_CLOSED
            - LISTING_EXPIRED
          in: query
          name: event_type
          type: string
        - description: Filter by actor ID
          in: query
          name: actor_id
          type: string
        - description: Filter events from this time (RFC3339 format)
          in: query
          name: time_from
          type: string
        - description: Filter events to this time (RFC3339 format)
          in: query
          name: time_to
          type: string
        - default: 1
          description: Page number
          in: query
          name: page
          type: integer
        - default: 50
          description: Items per page
          in: query
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Audit log retrieved successfully
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/marketplace.AuditLogResponse"
                type: object
        "400":
          description: Invalid request parameters
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "403":
          description: Insufficient privileges
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
      security:
        - BearerAuth: []
      summary: Get marketplace audit log (Admin)
      tags:
        - Admin
  /admin/v1/marketplace/bids/{id}:
    delete:
      consumes:
        - application/json
      description: Remove a fraudulent or policy-violating bid from the marketplace
      parameters:
        - description: Bid ID
          in: path
          name: id
          required: true
          type: string
        - description: Remove bid request
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/marketplace.RemoveBidRequest"
      produces:
        - application/json
      responses:
        "200":
          description: Bid removed successfully
          schema:
            $ref: "#/definitions/common.APIResponse"
        "400":
          description: Invalid request data
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "403":
          description: Insufficient privileges
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "404":
          description: Bid not found
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
      security:
        - BearerAuth: []
      summary: Remove a bid (Admin)
      tags:
        - Admin
  /admin/v1/marketplace/listings:
    get:
      consumes:
        - application/json
      description:
        Retrieve all marketplace listings with admin privileges - no visibility
        restrictions
      parameters:
        - description: Filter by status
          enum:
            - ACTIVE
            - CLOSED
            - EXPIRED
            - CANCELLED
          in: query
          name: status
          type: string
        - description: Filter by seller ID
          in: query
          name: seller_id
          type: string
        - description: Filter by organization ID
          in: query
          name: organization_id
          type: string
        - description: Filter by product ID
          in: query
          name: product_id
          type: string
        - description: Filter by visibility
          enum:
            - PRIVATE
            - PUBLIC
            - NETWORK
            - ORGANIZATION
          in: query
          name: visibility
          type: string
        - description: Filter by auction type
          enum:
            - OPEN
            - CLOSED
          in: query
          name: auction_type
          type: string
        - description: Filter by minimum price
          in: query
          name: min_price
          type: string
        - description: Filter by maximum price
          in: query
          name: max_price
          type: string
        - default: 1
          description: Page number
          in: query
          name: page
          type: integer
        - default: 20
          description: Items per page
          in: query
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Listings retrieved successfully
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/marketplace.ListingListResponse"
                type: object
        "400":
          description: Invalid request parameters
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "403":
          description: Insufficient privileges
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
      security:
        - BearerAuth: []
      summary: Get all marketplace listings (Admin)
      tags:
        - Admin
  /admin/v1/marketplace/listings/{id}/force-close:
    post:
      consumes:
        - application/json
      description: Force close a marketplace listing regardless of its current state
      parameters:
        - description: Listing ID
          in: path
          name: id
          required: true
          type: string
        - description: Force close request
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/marketplace.ForceCloseListingRequest"
      produces:
        - application/json
      responses:
        "200":
          description: Listing force closed successfully
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/marketplace.ListingResponse"
                type: object
        "400":
          description: Invalid request data
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "403":
          description: Insufficient privileges
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "404":
          description: Listing not found
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
      security:
        - BearerAuth: []
      summary: Force close a marketplace listing (Admin)
      tags:
        - Admin
  /admin/v1/marketplace/stats:
    get:
      consumes:
        - application/json
      description: Retrieve comprehensive marketplace statistics and analytics
      parameters:
        - default: 30d
          description: Time range for statistics
          enum:
            - 24h
            - 7d
            - 30d
            - 90d
            - 1y
          in: query
          name: time_range
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Statistics retrieved successfully
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  data:
                    $ref: "#/definitions/marketplace.MarketplaceStatsResponse"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "403":
          description: Insufficient privileges
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: "#/definitions/common.APIResponse"
              - properties:
                  error:
                    $ref: "#/definitions/common.APIError"
                type: object
      security:
        - BearerAuth: []
      summary: Get marketplace statistics (Admin)
      tags:
        - Admin
...
swagger: "2.0"